# [Inductor] Check if profiling before using record_function in CompiledFxGraph